refactor the structure
