+++ Python/Product/Profiling/Profiling/UserInputService.cs
@@ -28,7 +28,15 @@ namespace Microsoft.PythonTools.Profiling {
     /// Implements a service to collect user input for profiling and generate a <see cref="TargetCommand"/>.
     /// </summary>
     [Export(typeof(IUserInputService))]
-    internal class UserInputService : IUserInputService {
+    class UserInputService : IUserInputService {
+        private readonly CommandBuilder _commandBuilder;
+        private readonly UserDialog _userDialog;
+
+        [ImportingConstructor]
+        public UserInputService() {
+            _commandBuilder = new CommandBuilder();
+            _userDialog = new UserDialog();
+        }
 
         /// <summary>
         /// Collects user input and constructs a <see cref="TargetCommand"/> object.
@@ -37,26 +45,13 @@ namespace Microsoft.PythonTools.Profiling {
         /// A <see cref="TargetCommand"/> object based on user input, or <c>null</c> if canceled.
         /// </returns>
         public TargetCommand GetCommandFromUserInput() {
-            return ShowUserInputDialog();
-        }
-
-        /// <summary>
-        /// Displays the user input dialog and processes the input to generate a <see cref="TargetCommand"/>.
-        /// </summary>
-        /// <returns>
-        /// A <see cref="TargetCommand"/> object based on the user's input, or <c>null</c> if canceled or invalid input.
-        /// </returns>
-        private TargetCommand ShowUserInputDialog() {
             try {
                 var pythonProfilingPackage = PythonProfilingPackage.Instance;
                 var targetView = new ProfilingTargetView(pythonProfilingPackage);
-                var userInputDialog = new LaunchProfiling(pythonProfilingPackage, targetView);
 
-                var result = userInputDialog.ShowModal() ?? false;
-
-                if (result && targetView.IsValid) {
+                if (_userDialog.ShowDialog(targetView)) {
                     var target = targetView.GetTarget();
-                    return BuildCommandFromTarget(target);
+                    return _commandBuilder.BuildCommandFromTarget(target);
                 }
                 return null;
             } catch (Exception ex) {
@@ -65,144 +60,5 @@ namespace Microsoft.PythonTools.Profiling {
                 return null;
             }
         }
-
-        /// <summary>
-        /// Constructs a <see cref="TargetCommand"/> based on the provided profiling target.
-        /// </summary>
-        private TargetCommand BuildCommandFromTarget(ProfilingTarget target) {
-            try {
-                if (target == null) {
-                    return null;
-                }
-
-                var pythonProfilingPackage = PythonProfilingPackage.Instance;
-                var joinableTaskFactory = pythonProfilingPackage.JoinableTaskFactory;
-
-                TargetCommand command = null;
-
-                joinableTaskFactory.Run(async () => {
-                    await joinableTaskFactory.SwitchToMainThreadAsync();
-
-                    var name = target.GetProfilingName(pythonProfilingPackage, out var save);
-                    var explorer = await pythonProfilingPackage.ShowPerformanceExplorerAsync();
-                    var session = explorer.Sessions.AddTarget(target, name, save);
-
-                    command = GenerateCommand(target, session);
-
-                });
-
-                return command;
-            } catch (Exception ex) {
-                Debug.Fail($"Error building command: {ex.Message}");
-                throw;
-            }
-        }
-
-        /// <summary>
-        /// Generates a <see cref="TargetCommand"/> based on the profiling target and session details.
-        /// </summary>
-        private TargetCommand GenerateCommand(ProfilingTarget target, SessionNode session) {
-            var projectTarget = target.ProjectTarget;
-            var standaloneTarget = target.StandaloneTarget;
-
-            if (projectTarget != null) {
-                return BuildCommandForProjectTarget(projectTarget, session);
-            } else if (standaloneTarget != null) {
-                return BuildCommandForStandaloneTarget(standaloneTarget, session);
-            }
-            return null;
-        }
-
-        private TargetCommand BuildCommandForProjectTarget(ProjectTarget projectTarget, SessionNode session) {
-            var solution = PythonProfilingPackage.Instance.Solution;
-            var project = solution.EnumerateLoadedPythonProjects()
-                .SingleOrDefault(p => p.GetProjectIDGuidProperty() == projectTarget.TargetProject);
-
-            if (project == null) {
-                return null;
-            }
-
-            LaunchConfiguration config = null;
-            try {
-                config = project?.GetLaunchConfigurationOrThrow();
-            } catch (NoInterpretersException ex) {
-                PythonToolsPackage.OpenNoInterpretersHelpPage(session._serviceProvider, ex.HelpPage);
-                return null;
-            } catch (MissingInterpreterException ex) {
-                MessageBox.Show(ex.Message, Strings.ProductTitle);
-                return null;
-            } catch (IOException ex) {
-                MessageBox.Show(ex.Message, Strings.ProductTitle);
-                return null;
-            }
-            if (config == null) {
-                MessageBox.Show(Strings.ProjectInterpreterNotFound.FormatUI(project.GetNameProperty()), Strings.ProductTitle);
-                return null;
-            }
-
-            if (string.IsNullOrEmpty(config.ScriptName)) {
-                MessageBox.Show(Strings.NoProjectStartupFile, Strings.ProductTitle);
-                return null;
-            }
-
-            if (string.IsNullOrEmpty(config.WorkingDirectory) || config.WorkingDirectory == ".") {
-                config.WorkingDirectory = project.ProjectHome;
-                if (string.IsNullOrEmpty(config.WorkingDirectory)) {
-                    config.WorkingDirectory = Path.GetDirectoryName(config.ScriptName);
-                }
-            }
-
-            var pythonExePath = config.GetInterpreterPath();
-            var scriptPath = string.Join(" ", ProcessOutput.QuoteSingleArgument(config.ScriptName), config.ScriptArguments);
-            var workingDir = config.WorkingDirectory;
-            var envVars = session._serviceProvider.GetPythonToolsService().GetFullEnvironment(config)
-
-            var command = new TargetCommand {
-                PythonExePath = pythonExePath,
-                ScriptPath = scriptPath,
-                WorkingDir = workingDir,
-                Args = null,
-                EnvVars = envVars
-            };
-            return command;
-        }
-
-        private TargetCommand BuildCommandForStandaloneTarget(StandaloneTarget standaloneTarget, SessionNode session) {
-            if (standaloneTarget == null) {
-                return null;
-            }
-
-            LaunchConfiguration config = null;
-
-            if (standaloneTarget.InterpreterPath != null) {
-                config = new LaunchConfiguration(null);
-            }
-
-            if (standaloneTarget.PythonInterpreter != null) {
-                var registry = session._serviceProvider.GetComponentModel().GetService<IInterpreterRegistryService>();
-                var interpreter = registry.FindConfiguration(standaloneTarget.PythonInterpreter.Id);
-                if (interpreter == null) {
-                    return null;
-                }
-
-                config = new LaunchConfiguration(interpreter);
-            }
-
-            config.InterpreterPath = standaloneTarget.InterpreterPath;
-            config.ScriptName = standaloneTarget.Script;
-            config.ScriptArguments = standaloneTarget.Arguments;
-            config.WorkingDirectory = standaloneTarget.WorkingDirectory;
-
-            var envVars = session._serviceProvider.GetPythonToolsService().GetFullEnvironment(config);
-
-            return new TargetCommand {
-                PythonExePath = config.GetInterpreterPath(),
-                WorkingDir = standaloneTarget.WorkingDirectory,
-                ScriptPath = standaloneTarget.Script,
-                Args = standaloneTarget.Arguments,
-                EnvVars = envVars
-            };
-        }
-
     }
 }
